backend/requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
pydantic